Feature slug permissions (#18)
* Campo Slug en los permisos

* Agregue nuevo campo slug en la tabla permissions

* Cambiando el nombre de las rutas en el api de permissions

* Corregi los campos que se visualizan en el swagger de la interfaz Permissions

* Corregi las reglas del FormRequest de Permissions

* Un comentario para hacer pull

+++ app/Repositories/PermissionRepository.php
@@ -20,6 +20,6 @@ class PermissionRepository extends BaseRepository
     }
 
     public function deleteRoleHasPermission($id) {
-        DB::table('role_has_permissions')->where('permission_id', $id)->delete();
+        DB::connection('tenant')->table('role_has_permissions')->where('permission_id', $id)->delete();
     }
 }
